backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
python-docx==1.1.2
pypdf==5.0.1
reportlab==4.2.5
python-multipart==0.0.9
